Modify chekcout counter to count total quantity from database instead of counter
