+++ routes/hello.json
@@ -1,0 +1,110 @@
+{
+  "returnObject": {
+    "items": [
+      {
+        "id": 1,
+        "name": "Good Potatoes",
+        "category": "appetizers",
+        "description": "These are potatoes. The good kind.",
+        "price": 699,
+        "img_url": "https://www.allrecipes.com/thmb/XvgVQJcbAKPgYqhkjlpekJEOdtg=/750x0/filters:no_upscale():max_bytes(150000):strip_icc():format(webp)/Best-Potatoes-Youll-Ever-Taste-a827fec107d44119b8c700f4091408ee.jpg",
+        "restaurant_id": 1
+      },
+      {
+        "id": 2,
+        "name": "Lasagna",
+        "category": "mains",
+        "description": "A tasty pickle lasagna",
+        "price": 1600,
+        "img_url": "https://www.thewholesomedish.com/wp-content/uploads/2018/07/Best-Lasagna-550-500x375.jpg",
+        "restaurant_id": 1
+      },
+      {
+        "id": 3,
+        "name": "Garlic Bread",
+        "category": "appetizers",
+        "description": "Calabrese bread, garlic butter and mozzarella",
+        "price": 600,
+        "img_url": "https://food.fnr.sndimg.com/content/dam/images/food/fullset/2015/5/28/2/TM1A14F_Garlic-Bread_s4x3.jpg.rend.hgtvcom.616.462.suffix/1433523400627.jpeg",
+        "restaurant_id": 1
+      },
+      {
+        "id": 4,
+        "name": "Bruschetta",
+        "category": "appetizers",
+        "description": "Garlic bread, mozzarella, diced tomatoes, garlic and olive oil",
+        "price": 750,
+        "img_url": "https://www.allrecipes.com/thmb/ZSIy3gJcPu2i-jNRwNvcYi2zEdw=/1500x0/filters:no_upscale():max_bytes(150000):strip_icc()/54165-Balsamic-Bruschetta-mfs_002-5a7881578a854bb187327708ee78503c.jpg",
+        "restaurant_id": 1
+      },
+      {
+        "id": 5,
+        "name": "Nonas famous Ravioli",
+        "category": "mains",
+        "description": "Spinach & Ricotta w/ pomodoro sauce",
+        "price": 1850,
+        "img_url": "https://timeincsecure-a.akamaihd.net/rtmp_uds/429048911/201806/2860/429048911_5798102596001_5798101558001-vs.jpg?pubId=1260616200&videoId=5793422078001",
+        "restaurant_id": 1
+      },
+      {
+        "id": 6,
+        "name": "Gnocchi Marinara",
+        "category": "mains",
+        "description": "potato dumplings sauteed with tomato sauce",
+        "price": 1700,
+        "img_url": "https://www.godairyfree.org/wp-content/uploads/2016/07/pics-Dairy-Free-Creamy-Tomato-Basil-Sauce-feature.jpg",
+        "restaurant_id": 1
+      },
+      {
+        "id": 7,
+        "name": "Penne Bolognese",
+        "category": "mains",
+        "description": "diagonal ridged tubes sauteed with meat sauce",
+        "price": 1700,
+        "img_url": "https://izzycooking.com/wp-content/uploads/2022/07/Penne-Bolognese-thumbnail.jpg",
+        "restaurant_id": 1
+      },
+      {
+        "id": 8,
+        "name": "Cannoli",
+        "category": "snacks",
+        "description": "filled with a sweet ricotta and creme cheese filling,",
+        "price": 125,
+        "img_url": "https://www.cookiedoughandovenmitt.com/wp-content/uploads/2014/09/Cannoli-Cream-Square.jpg",
+        "restaurant_id": 1
+      },
+      {
+        "id": 9,
+        "name": "Gelato",
+        "category": "snacks",
+        "description": "light, unbelievably smooth frozen dessert",
+        "price": 450,
+        "img_url": "https://www.allrecipes.com/thmb/I8XVYqnflHCOotWRhlFWCPAftpg=/1500x0/filters:no_upscale():max_bytes(150000):strip_icc()/4517634-gelato-Buckwheat-Queen-4x3-1-f5f732166ec64620b2a438391cbd0583.jpg",
+        "restaurant_id": 1
+      },
+      {
+        "id": 10,
+        "name": "Espresso",
+        "category": "drinks",
+        "description": "rich and bold",
+        "price": 350,
+        "img_url": "https://www.thespruceeats.com/thmb/HJrjMfXdLGHbgMhnM0fMkDx9XPQ=/1500x0/filters:no_upscale():max_bytes(150000):strip_icc()/what-is-espresso-765702-hero-03_cropped-ffbc0c7cf45a46ff846843040c8f370c.jpg",
+        "restaurant_id": 1
+      },
+      {
+        "id": 11,
+        "name": "Italian Soda",
+        "category": "drinks",
+        "description": "available in: Cherry, Blueberry or Pineapple",
+        "price": 300,
+        "img_url": "https://i2.wp.com/lmld.org/wp-content/uploads/2018/02/Italian-Cream-Sodas-3.jpg",
+        "restaurant_id": 1
+      }
+    ],
+    "orderItems": [
+      { "name": "Penne Bolognese", "price": 1700, "quantity": 1 },
+      { "name": "Nonas famous Ravioli", "price": 1850, "quantity": 1 }
+    ],
+    "totalQuantity": 2
+  }
+}
